backend/engine/src/main/java/com/joma/studies/QueryParser.java
package com.joma.studies;

import org.apache.lucene.analysis.Analyzer;
import org.apache.lucene.analysis.TokenStream;
import org.apache.lucene.analysis.standard.StandardAnalyzer;
import org.apache.lucene.analysis.tokenattributes.CharTermAttribute;

import java.io.IOException;
import java.io.StringReader;
import java.util.HashMap;
import java.util.Map;

public class QueryParser {

    public Map<String, Integer> parse(String query) {
        HashMap<String, Integer> result = new HashMap<>();

        Analyzer analyzer = new StandardAnalyzer();
        TokenStream tokenStream = analyzer.tokenStream("field", new StringReader(query));
        CharTermAttribute charTermAttribute = tokenStream.addAttribute(CharTermAttribute.class);

        try {
            tokenStream.reset();
            while (tokenStream.incrementToken()) {
                String term = charTermAttribute.toString();
                if (result.containsKey(term)) {
                    Integer value = result.get(term);
                    result.replace(term, value, ++value);
                } else {
                    result.put(charTermAttribute.toString(), 1);
                }
            }
            tokenStream.end();
        } catch (IOException e) {
            e.printStackTrace();
        } finally {
            try {
                tokenStream.close();
            } catch (IOException e) {
                e.printStackTrace();
            }
        }

        return result;
    }
}
